Prioritize RPC-related types in schema reference
This changes the order of the schema reference page such that
RPC-related types are listed first and common types are listed last.
This also adds logic to ensure that, for a given RPC method, the
relevant types are ordered as follows: `*Request` type, `*RequestParams`
type, `*Result` type, `*Notification` type, `*NotificationParams` type,
and then any other relevant types in alphabetical order.