Generate schema reference page based on template
This commit modifies the TypeDoc plugin to generate the schema reference
page based on a template `schema.mdx` file instead of based solely on
the content of `schema.ts`.  The template file uses placeholders to
indicate where categories of types should be rendered.  For example, the
``{/* @category `tools/call` */}`` placeholder will be replaced by the
documentation for all types in the `` `tools/call` `` category.

This change gives us more control over headings and category order, and
allows us to easily add surrounding prose.