Hide `@TJS-type` tag artifacts in `schema.mdx`
Previously, `@TJS-type` tags were not recognized by TypeDoc because of
the hyphen in their name.  However, [TypeDoc v0.28.8][] supports such
tags, allowing them to be listed in the `excludeTags` option.

This commit upgrades TypeDoc to the latest version, and adds `@TJS-type`
tags to the `excludeTags` configuration option, preventing their
appearance in `schema.mdx`.

[TypeDoc v0.28.8]: https://github.com/TypeStrong/typedoc/releases/tag/v0.28.8